Prepare v2020.07 for Orange Pi R1 Plus

+++ drivers/spi/atcspi200_spi.c
@@ -6,8 +6,9 @@
  * Author: Rick Chen (rick@andestech.com)
  */
 
-#include <clk.h>
 #include <common.h>
+#include <clk.h>
+#include <log.h>
 #include <malloc.h>
 #include <spi.h>
 #include <asm/io.h>
@@ -230,14 +231,14 @@ static int __atcspi200_spi_xfer(struct nds_spi_slave *ns,
 			break;
 		}
 		if (data_out)
-			debug("spi_xfer: data_out %08X(%p) data_in %08X(%p) data_len %u\n",
+			debug("spi_xfer: data_out %08X(%p) data_in %08X(%p) data_len %lu\n",
 			      *(uint *)data_out, data_out, *(uint *)data_in,
 			      data_in, data_len);
 		num_chunks = DIV_ROUND_UP(data_len, max_tran_len);
 		din = data_in;
 		dout = data_out;
 		while (num_chunks--) {
-			tran_len = min(data_len, (size_t)max_tran_len);
+			tran_len = min((size_t)data_len, (size_t)max_tran_len);
 			ns->tran_len = tran_len;
 			num_blks = DIV_ROUND_UP(tran_len , CHUNK_SIZE);
 			num_bytes = (tran_len) % CHUNK_SIZE;
